+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- strip `kubectl.kubernetes.io/last-applied-configuration` annotation
+
 ## [0.2.0] - 2021-05-14
 ### Changed
 - close active editor on create